division

+++ src/main/java/com/company/Calculator.java
@@ -13,9 +13,9 @@ public class Calculator {
         return ((long) a * b);
     }
 
-    public int division(int a, int b) {
+    public double division(int a, int b) {
         if (b != 0) {
-            return (a / b);
+            return ((double) a / b);
         }
         else {
             System.out.println("Нельзя делить на ноль");
